fixed decimal transfers being blocked by trigger

+++ migrations/20190330002201-CreateAccountsTriggerFunctionn.sql
@@ -7,7 +7,7 @@ CREATE OR REPLACE FUNCTION check_if_account_balanced()
 RETURNS TRIGGER
 AS $$
 DECLARE
-  total integer;
+  total decimal;
 BEGIN
   total := (SELECT COALESCE(SUM(CASE WHEN direction = 'outgoing' THEN amount * -1 ELSE amount END), 0) FROM payments WHERE account_id = NEW.id);
   IF (total != NEW.balance) THEN
